bzr revid: p_christ@hol.gr-20110117094021-rbjm55dip17tnuy4

+++ bin/addons/base/res/res_user.py
@@ -3,6 +3,7 @@
 #
 #    OpenERP, Open Source Management Solution
 #    Copyright (C) 2004-2009 Tiny SPRL (<http://tiny.be>).
+#    Copyright (C) 2010-2011 OpenERP s.a. (<http://openerp.com>).
 #
 #    This program is free software: you can redistribute it and/or modify
 #    it under the terms of the GNU Affero General Public License as
@@ -28,16 +29,16 @@ import pooler
 from tools.translate import _
 from service import security
 import netsvc
-import time
 
 class groups(osv.osv):
     _name = "res.groups"
     _order = 'name'
+    _description = "Access Groups"
     _columns = {
         'name': fields.char('Group Name', size=64, required=True),
         'model_access': fields.one2many('ir.model.access', 'group_id', 'Access Controls'),
         'rule_groups': fields.many2many('ir.rule', 'rule_group_rel',
-            'group_id', 'rule_group_id', 'Rules', domain="[('global', '<>', True)]"),
+            'group_id', 'rule_group_id', 'Rules', domain=[('global', '=', False)]),
         'menu_access': fields.many2many('ir.ui.menu', 'ir_ui_menu_group_rel', 'gid', 'menu_id', 'Access Menu'),
         'comment' : fields.text('Comment',size=250),
     }
@@ -45,7 +46,7 @@ class groups(osv.osv):
         ('name_uniq', 'unique (name)', 'The name of the group must be unique !')
     ]
 
-    def copy(self, cr, uid, id, default=None, context={}):
+    def copy(self, cr, uid, id, default=None, context=None):
         group_name = self.read(cr, uid, [id], ['name'])[0]['name']
         default.update({'name': _('%s (copy)')%group_name})
         return super(groups, self).copy(cr, uid, id, default, context)
@@ -82,20 +83,21 @@ class groups(osv.osv):
 
 groups()
 
-def _lang_get(self, cr, uid, context={}):
+def _lang_get(self, cr, uid, context=None):
     obj = self.pool.get('res.lang')
     ids = obj.search(cr, uid, [('translatable','=',True)])
-    res = obj.read(cr, uid, ids, ['code', 'name'], context)
+    res = obj.read(cr, uid, ids, ['code', 'name'], context=context)
     res = [(r['code'], r['name']) for r in res]
     return res
 
-def _tz_get(self,cr,uid, context={}):
+def _tz_get(self,cr,uid, context=None):
     return [(x, x) for x in pytz.all_timezones]
 
 class users(osv.osv):
     __admin_ids = {}
     _uid_cache = {}
     _name = "res.users"
+    _order = 'name'
 
     WELCOME_MAIL_SUBJECT = u"Welcome to OpenERP"
     WELCOME_MAIL_BODY = u"An OpenERP account has been created for you, "\
@@ -192,6 +194,18 @@ class users(osv.osv):
                 self.write(cr, uid, ids, {'address_id': address_id}, context)
         return True
 
+    def _set_new_password(self, cr, uid, id, name, value, args, context=None):
+        if value is False:
+            # Do not update the password if no value is provided, ignore silently.
+            # For example web client submits False values for all empty fields.
+            return
+        if uid == id:
+            # To change their own password users must use the client-specific change password wizard,
+            # so that the new password is immediately used for further RPC requests, otherwise the user
+            # will face unexpected 'Access Denied' exceptions.
+            raise osv.except_osv(_('Operation Canceled'), _('Please use the change password wizard (in User Preferences or User menu) to change your own password.'))
+        self.write(cr, uid, id, {'password': value})
+
     _columns = {
         'name': fields.char('User Name', size=64, required=True, select=True,
                             help="The new user's real name, used for searching"
@@ -199,6 +213,10 @@ class users(osv.osv):
         'login': fields.char('Login', size=64, required=True,
                              help="Used to log into the system"),
         'password': fields.char('Password', size=64, invisible=True, help="Keep empty if you don't want the user to be able to connect on the system."),
+        'new_password': fields.function(lambda *a:'', method=True, type='char', size=64,
+                                fnct_inv=_set_new_password,
+                                string='Change password', help="Only specify a value if you want to change the user password. "
+                                "This user will have to logout and login again!"),
         'email': fields.char('E-mail', size=64,
             help='If an email is provided, the user will be sent a message '
                  'welcoming him.\n\nWarning: if "email_from" and "smtp_server"'
@@ -234,9 +252,10 @@ class users(osv.osv):
 
     def on_change_company_id(self, cr, uid, ids, company_id):
         return {
-            'value': {
-                'warning' : _("Please keep in mind that data currently displayed may not be relevant after switching to another company. If you have unsaved changes, please make sure to save and close the forms before switching to a different company (you can click on Cancel now)"),
-            }
+                'warning' : {
+                    'title': _("Company Switch Warning"),
+                    'message': _("Please keep in mind that documents currently displayed may not be relevant after switching to another company. If you have unsaved changes, please make sure to save and close all forms before switching to a different company. (You can click on Cancel in the User Preferences now)"),
+                }
         }
 
     def read(self,cr, uid, ids, fields=None, context=None, load='_classic_read'):
@@ -263,7 +282,7 @@ class users(osv.osv):
     ]
 
     _sql_constraints = [
-        ('login_key', 'UNIQUE (login)',  _('You can not have two users with the same login !'))
+        ('login_key', 'UNIQUE (login)',  'You can not have two users with the same login !')
     ]
 
     def _get_email_from(self, cr, uid, ids, context=None):
@@ -296,8 +315,14 @@ class users(osv.osv):
         return False
 
     def _get_menu(self,cr, uid, context=None):
-        ids = self.pool.get('ir.actions.act_window').search(cr, uid, [('usage','=','menu')], context=context)
-        return ids and ids[0] or False
+        dataobj = self.pool.get('ir.model.data')
+        try:
+            model, res_id = dataobj.get_object_reference(cr, uid, 'base', 'action_menu_admin')
+            if model != 'ir.actions.act_window':
+                return False
+            return res_id
+        except ValueError:
+            return False
 
     def _get_group(self,cr, uid, context=None):
         dataobj = self.pool.get('ir.model.data')
@@ -313,9 +338,9 @@ class users(osv.osv):
         return result
 
     _defaults = {
-        'password' : lambda *a : '',
-        'context_lang': lambda *args: 'en_US',
-        'active' : lambda *a: True,
+        'password' : '',
+        'context_lang': 'en_US',
+        'active' : True,
         'menu_id': _get_menu,
         'company_id': _get_company,
         'company_ids': _get_companies,
@@ -351,12 +376,22 @@ class users(osv.osv):
         self.pool.get('ir.model.access').call_cache_clearing_methods(cr)
         clear = partial(self.pool.get('ir.rule').clear_cache, cr)
         map(clear, ids)
+        db = cr.dbname
+        if db in self._uid_cache:
+            for id in ids:
+                if id in self._uid_cache[db]:
+                    del self._uid_cache[db][id]
 
         return res
 
     def unlink(self, cr, uid, ids, context=None):
         if 1 in ids:
             raise osv.except_osv(_('Can not remove root user!'), _('You can not remove the admin user as it is used internally for resources created by OpenERP (updates, module installation, ...)'))
+        db = cr.dbname
+        if db in self._uid_cache:
+            for id in ids:
+                if id in self._uid_cache[db]:
+                    del self._uid_cache[db][id]
         return super(users, self).unlink(cr, uid, ids, context=context)
 
     def name_search(self, cr, user, name='', args=None, operator='ilike', context=None, limit=100):
@@ -371,16 +406,17 @@ class users(osv.osv):
             ids = self.search(cr, user, [('name',operator,name)]+ args, limit=limit)
         return self.name_get(cr, user, ids)
 
-    def copy(self, cr, uid, id, default=None, context={}):
+    def copy(self, cr, uid, id, default=None, context=None):
         user2copy = self.read(cr, uid, [id], ['login','name'])[0]
         if default is None:
             default = {}
         copy_pattern = _("%s (copy)")
-        default.update(login=(copy_pattern % user2copy['login']),
+        copydef = dict(login=(copy_pattern % user2copy['login']),
                        name=(copy_pattern % user2copy['name']),
                        address_id=False, # avoid sharing the address of the copied user!
                        )
-        return super(users, self).copy(cr, uid, id, default, context)
+        copydef.update(default)
+        return super(users, self).copy(cr, uid, id, copydef, context)
 
     def context_get(self, cr, uid, context=None):
         user = self.browse(cr, uid, uid, context)
@@ -393,25 +429,28 @@ class users(osv.osv):
                 result[k[8:]] = res or False
         return result
 
-    def action_get(self, cr, uid, context={}):
+    def action_get(self, cr, uid, context=None):
         dataobj = self.pool.get('ir.model.data')
         data_id = dataobj._get_id(cr, 1, 'base', 'action_res_users_my')
-        return dataobj.browse(cr, uid, data_id, context).res_id
+        return dataobj.browse(cr, uid, data_id, context=context).res_id
 
 
     def login(self, db, login, password):
         if not password:
             return False
         cr = pooler.get_db(db).cursor()
-        cr.execute('select id from res_users where login=%s and password=%s and active', (tools.ustr(login), tools.ustr(password)))
-        res = cr.fetchone()
-        result = False
-        if res:
-            cr.execute("update res_users set date=%s where id=%s", (time.strftime('%Y-%m-%d %H:%M:%S'),res[0]))
+        try:
+            cr.execute('UPDATE res_users SET date=now() WHERE login=%s AND password=%s AND active RETURNING id',
+                    (tools.ustr(login), tools.ustr(password)))
+            res = cr.fetchone()
             cr.commit()
-            result = res[0]
-        cr.close()
-        return result
+            if res:
+                return res[0]
+            else:
+                return False
+        finally:
+            cr.close()
+
     def check_super(self, passwd):
         if passwd == tools.config['admin_passwd']:
             return True
@@ -421,33 +460,51 @@ class users(osv.osv):
     def check(self, db, uid, passwd):
         if not passwd:
             return False
-        cached_pass = self._uid_cache.get(db, {}).get(uid)
-        if (cached_pass is not None) and cached_pass == passwd:
+        if self._uid_cache.get(db, {}).get(uid) == passwd:
             return True
         cr = pooler.get_db(db).cursor()
-        cr.execute('select count(1) from res_users where id=%s and password=%s and active=%s', (int(uid), passwd, True))
-        res = cr.fetchone()[0]
-        cr.close()
-        if not bool(res):
-            raise security.ExceptionNoTb('AccessDenied')
-        if res:
-            if self._uid_cache.has_key(db):
-                ulist = self._uid_cache[db]
-                ulist[uid] = passwd
-            else:
-                self._uid_cache[db] = {uid:passwd}
-        return bool(res)
+        try:
+            cr.execute('SELECT COUNT(1) FROM res_users WHERE id=%s AND password=%s AND active=%s',
+                        (int(uid), passwd, True))
+            res = cr.fetchone()[0]
+            if not bool(res):
+                raise security.ExceptionNoTb('AccessDenied')
+            if res:
+                if self._uid_cache.has_key(db):
+                    ulist = self._uid_cache[db]
+                    ulist[uid] = passwd
+                else:
+                    self._uid_cache[db] = {uid:passwd}
+            return bool(res)
+        finally:
+            cr.close()
 
     def access(self, db, uid, passwd, sec_level, ids):
         if not passwd:
             return False
         cr = pooler.get_db(db).cursor()
-        cr.execute('select id from res_users where id=%s and password=%s', (uid, passwd))
-        res = cr.fetchone()
-        cr.close()
-        if not res:
-            raise security.ExceptionNoTb('Bad username or password')
-        return res[0]
+        try:
+            cr.execute('SELECT id FROM res_users WHERE id=%s AND password=%s', (uid, passwd))
+            res = cr.fetchone()
+            if not res:
+                raise security.ExceptionNoTb('Bad username or password')
+            return res[0]
+        finally:
+            cr.close()
+
+    def change_password(self, cr, uid, old_passwd, new_passwd, context=None):
+        """Change current user password. Old password must be provided explicitly
+        to prevent hijacking an existing user session, or for cases where the cleartext
+        password is not used to authenticate requests.
+
+        :return: True
+        :raise: security.ExceptionNoTb when old password is wrong
+        :raise: except_osv when new password is not set or empty
+        """
+        self.check(cr.dbname, uid, old_passwd)
+        if new_passwd:
+            return self.write(cr, uid, uid, {'password': new_passwd})
+        raise osv.except_osv(_('Warning!'), _("Setting empty passwords is not allowed for security reasons!"))
 
 users()
 
@@ -510,9 +567,19 @@ class groups2(osv.osv): ##FIXME: Is there a reason to inherit this object ?
     }
 
     def unlink(self, cr, uid, ids, context=None):
+        group_users = []
         for record in self.read(cr, uid, ids, ['users'], context=context):
             if record['users']:
-                raise osv.except_osv(_('Warning !'), _('Make sure you have no users linked with the group(s)!'))
+                group_users.extend(record['users'])
+
+        if group_users:
+            user_names = [user.name for user in self.pool.get('res.users').browse(cr, uid, group_users, context=context)]
+            if len(user_names) >= 5:
+                user_names = user_names[:5]
+                user_names += '...'
+            raise osv.except_osv(_('Warning !'),
+                        _('Group(s) cannot be deleted, because some user(s) still belong to them: %s !') % \
+                            ', '.join(user_names))
         return super(groups2, self).unlink(cr, uid, ids, context=context)
 
 groups2()
@@ -538,4 +605,3 @@ class res_config_view(osv.osv_memory):
 res_config_view()
 
 # vim:expandtab:smartindent:tabstop=4:softtabstop=4:shiftwidth=4:
-
